delete confirmation

+++ public/javascripts/delete_user.js
@@ -2,9 +2,12 @@ $(clickHandler);
 
 function clickHandler() {
   $(".deleteButton").click(function() {
+
+    if (confirm('Are you sure?')) {
     const uid = $(this).attr("data-uid");
     console.log("delete: ", uid);
     deleteUser(uid);
+    } else {}
   });
 }
